+++ utils/jest.config.ts
@@ -1,0 +1,17 @@
+import type { Config } from 'jest';
+
+const config: Config = {
+  preset: 'ts-jest',
+  testEnvironment: 'node',
+  testMatch: ['<rootDir>/tests/**/*.test.ts'],
+  moduleFileExtensions: ['ts', 'js'],
+  verbose: true,
+  collectCoverage: true,
+  coverageDirectory: 'coverage',
+  transform: {
+    '^.+\\.(ts|js)$': 'babel-jest'
+  },
+  transformIgnorePatterns: []
+};
+
+export default config;